Import changes from GUWA project. - Derive bp_gaussian beam pattern from bp_solid. - Add -fPIC to USMLUse.cmake file - Change platform IDs from int to uint64_t to better support JSE/GRID ids. - Ignore -Wmaybe-uninitialized from boost::geomentry in eigenverb_collection.h - Add ability to read USML_DATA_DIR from environment variables in ocean_utils::make_basic() function. - Add is_acoustic_target flag to platform_model so that it can be skipped in wavefront generation. - Fix index calculation in rvbts_collection.cc file. - Add ability to read time_series data from rvbts_collection. - Add treverb parameter to rvbts_generator as replacement for fsample. - Add ability to reset_src_beams() in sensor_model. - Get rid of fsample and fcenter in sensor_model. - Fix the way that wavefront generators are computed. - Fix the way that thead_controller is reset. - Minor clang-tidy fixes.

+++ waveq3d/test/reverb_test.cc
@@ -100,8 +100,8 @@ BOOST_AUTO_TEST_CASE(dead_reckon) {
 
     // test ability to write target id values to disk
 
-    matrix<int> targetIDs(1, 2);
-    for (int n = 0; n < targetIDs.size2(); n++) {
+    matrix<uint64_t> targetIDs(1, 2);
+    for (size_t n = 0; n < targetIDs.size2(); n++) {
         targetIDs(0, n) = n + 2;
     }
